requirements.txt
asgiref==3.2.10
Django==3.0
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.1
Markdown==3.2.2
django-filter==2.3.0
psycopg2-binary==2.8.5
django-allauth==0.42.0
django-rest-auth==0.9.5
django-phonenumber-field==5.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.6.0
